refactor data folder
