+++ scripts/bert/finetune_classifier.py
@@ -50,8 +50,9 @@ from gluonnlp.model import get_bert_model
 from gluonnlp.data import BERTTokenizer
 
 from model.classification import BERTClassifier, BERTRegression
-from dataset import MRPCTask, QQPTask, RTETask, STSBTask, \
-    QNLITask, CoLATask, MNLITask, WNLITask, SSTTask, XNLITask, BERTDatasetTransform
+from data.classification import MRPCTask, QQPTask, RTETask, STSBTask, SSTTask
+from data.classification import QNLITask, CoLATask, MNLITask, WNLITask, XNLITask
+from data.transform import BERTDatasetTransform
 
 tasks = {
     'MRPC': MRPCTask(),
