+++ scripts/bert/index.rst
@@ -81,7 +81,7 @@ For all model settings above, we set learing rate = 2e-5 and optimizer = bertada
 
 .. code-block:: console
 
-    $ BAIDU_ERNIE_DATA_DIR=baidu_ernie_data python finetune_classifier.py --seed 6 --task_name XNLI --batch_size 32 --optimizer bertadam --epochs 4 --lr 2e-5 --bert_dataset wiki_cn_cased --gpu 0
+    $ python finetune_classifier.py --seed 6 --task_name XNLI --batch_size 32 --epochs 4 --lr 2e-5 --bert_dataset wiki_cn_cased --gpu 0
 
 Some other tasks can be modeled with `--task_name` parameter.
